+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## master
+## 8.1.1 Dec 5, 2021
 
 Contributed: